bca-agent/pyproject.toml
[project]
name = "bca-agent"
version = "0.1.0"
description = "Banker Connections Agent - Clear CUID POC"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.132.0",
    "uvicorn[standard]>=0.27.0",
    "google-adk>=1.25.1",
    "google-genai>=1.0.0",
    "httpx>=0.26.0",
    "pydantic>=2.6.0",
    "pydantic-settings>=2.1.0",
    "python-dotenv>=1.0.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"


[tool.hatch.build.targets.wheel]
packages = ["app/"]
